Move port definition(s) to include/hamlib/port.h
Add some ghost structs and typedefs
Make example.c compile on gcc15 and with new includes
