+++ tests/example.c
@@ -11,11 +11,13 @@
 #include <stdio.h>
 #include <string.h>
 #include <hamlib/rig.h>
+#include <hamlib/port.h>
+#include <hamlib/rig_state.h>
 #include <hamlib/riglist.h>
 #include "sprintflst.h"
 #include <hamlib/rotator.h>
 
-#if 0
+#if 1
 #define MODEL RIG_MODEL_DUMMY
 #define PATH "/dev/ttyUSB0"
 #define BAUD 19200
@@ -31,14 +33,14 @@ int main()
     char *info_buf;
     freq_t freq;
     value_t rawstrength, power, strength;
-    float s_meter, rig_raw2val();
+    float s_meter, rig_raw2val(int i, cal_table_t *t);
     int status, retcode;
     unsigned int mwpower;
     rmode_t mode;
     pbwidth_t width;
 
     /* Set verbosity level */
-    rig_set_debug(RIG_DEBUG_TRACE);       // errors only
+    rig_set_debug(RIG_DEBUG_TRACE);       // Lots of output
 
     /* Instantiate a rig */
     my_rig = rig_init(MODEL); // your rig model.
@@ -123,7 +125,8 @@ int main()
 
     printf("LEVEL_STRENGTH returns %d\n", strength.i);
 
-    const freq_range_t *range = rig_get_range(&my_rig->state.rx_range_list[0],
+    const struct rig_state *my_rs = HAMLIB_STATE(my_rig);
+    const freq_range_t *range = rig_get_range(&my_rs->rx_range_list[0],
                                 14074000, RIG_MODE_USB);
 
     if (status != RIG_OK) { rig_debug(RIG_DEBUG_ERR, "%s: error rig_get_ragne: %s\n", __func__, rigerror(status)); }
@@ -131,7 +134,7 @@ int main()
     if (range)
     {
         char vfolist[256];
-        rig_sprintf_vfo(vfolist, sizeof(vfolist), my_rig->state.vfo_list);
+        rig_sprintf_vfo(vfolist, sizeof(vfolist), my_rs->vfo_list);
         printf("Range start=%"PRIfreq", end=%"PRIfreq", low_power=%d, high_power=%d, vfos=%s\n",
                range->startf, range->endf, range->low_power, range->high_power, vfolist);
     }
